refactor planproject
now all entries are directly linked to the project and the tasks only exist as a list.
